refactor: centralize common structs
